Update rendering logic
